feat: combine all components, clean up completed task files, and update original files
