--- src/components/TaskFolder/TaskAnnouncementBanner.vue
@@ -1,12 +1,0 @@
-<template>
-  <div class="container-fluid text-center bg-warning ">
-    <p class="text-dark p-2">
-      Are you a university or school student looking for a tutoring center?
-      <a href="#" class="fw-bold text-decoration-underline">Talk to us</a>
-    </p>
-  </div>
-</template>
-
-<script setup></script>
-
-<style lang="scss" scoped></style>
