--- src/components/TaskFolder/TaskTopNav.vue
@@ -1,31 +1,0 @@
-<template>
-  <div class="bg-light py-2">
-    <div class="container">
-      <div class="row align-items-center text-center text-lg-start gy-2">
-        <!-- Contact Info -->
-        <div class="col-12 col-lg-4 d-flex align-items-center justify-content-center justify-content-lg-start gap-2">
-          <i class="fa-solid fa-phone"></i>
-          <p class="mb-0">+234 708 893 2240, +234 816 230 0241</p>
-        </div>
-
-        <!-- Location Info -->
-        <div class="col-12 col-lg-5 align-items-center justify-content-center justify-content-lg-start gap-2">
-         
-          <p class="mb-0"> <i class="fa-solid fa-location-dot"></i> 2, Awoyelu Street, Jogbo, off Ibadan Garage, Ijebu-Ode, Ogun State, Nigeria</p>
-        </div>
-
-        <!-- Social Icons -->
-        <div class="col-12 col-lg-3 d-flex align-items-center justify-content-center justify-content-lg-end gap-3">
-          <i class="fa-brands fa-facebook"></i>
-          <i class="fa-brands fa-twitter"></i>
-          <i class="fa-brands fa-instagram"></i>
-          <i class="fa-brands fa-linkedin"></i>
-        </div>
-      </div>
-    </div>
-  </div>
-</template>
-
-<script setup></script>
-
-<style scoped></style>
