+++ src/components/navigation/AnnouncementBanner.vue
@@ -1,6 +1,9 @@
 <template>
-  <div>
-    <h2>Announcement Banner</h2>
+  <div class="container-fluid text-center bg-warning">
+    <p class="text-dark p-2">
+      Are you a university or school student looking for a tutoring center?
+      <a href="#" class="fw-bold text-decoration-underline">Talk to us</a>
+    </p>
   </div>
 </template>
