+++ src/router/index.js
@@ -1,52 +1,24 @@
 import { createRouter, createWebHistory } from "vue-router";
 import Home from "@/views/Home.vue";
 import About from "@/views/About.vue";
-import TaskAnnouncementBanner from "@/components/TaskFolder/TaskAnnouncementBanner.vue";
-import TaskTopNav from "@/components/TaskFolder/TaskTopNav.vue";
-import TaskMainNav from "../components/TaskFolder/TaskMainNav.vue";
-import TaskHeroSection from "../components/TaskFolder/TaskHeroSection.vue";
-import TaskFooter from "../components/TaskFolder/TaskFooter.vue";
-import TaskLogin from "../components/TaskFolder/TaskLogin.vue";
 import TaskRegister from "../components/TaskFolder/TaskRegister.vue";
 import TaskAdditionalInfoForm from "../components/TaskFolder/TaskAdditionalInfoForm.vue";
 import TaskProfilePhotoUpload from "../components/TaskFolder/TaskProfilePhotoUpload.vue";
 import TaskPersonalDetailsForm from "../components/TaskFolder/TaskPersonalDetailsForm.vue";
 import TaskPaymentForm from "../components/TaskFolder/TaskPaymentForm.vue";
 import TaskStepIndicator from "../components/TaskFolder/TaskStepIndicator.vue";
+import notice from "../views/notice.vue";
+import Login from "../views/Login.vue";
 
 const routes = [
   { path: "/", name: "Home", component: Home },
   { path: "/about", name: "About", component: About },
   {
-    path: "/taskannouncementbanner",
-    name: "TaskAnnouncementBanner",
-    component: TaskAnnouncementBanner,
+    path: "/login",
+    name: "Login",
+    component: Login,
   },
-  {
-    path: "/tasktopnav",
-    name: "TaskTopNav",
-    component: TaskTopNav,
-  },
-  {
-    path: "/taskmainnav",
-    name: "TaskMainNav",
-    component: TaskMainNav,
-  },
-  {
-    path: "/taskHeroSection",
-    name: "TaskHeroSection",
-    component: TaskHeroSection,
-  },
-  {
-    path: "/taskLogin",
-    name: "TaskLogin",
-    component: TaskLogin,
-  },
-  {
-    path: "/taskFooter",
-    name: "TaskFooter",
-    component: TaskFooter,
-  },
+
   {
     path: "/taskRegister",
     name: "TaskRegister",
@@ -76,6 +48,11 @@ const routes = [
     path: "/taskStepIndicator",
     name: "TaskStepIndicator",
     component: TaskStepIndicator,
+  },
+  {
+    path: "/notice",
+    name: "notice",
+    component: notice,
   },
 ];
